+++ packages/plugin-button/src/constants.js
@@ -1,0 +1,20 @@
+export const BUTTON_TYPE = 'wix-draft-plugin-button';
+export const ALIGN_CENTER = 'center';
+export const settingsTabValue = 'settings';
+export const designTabValue = 'design';
+export const BUTTON_TYPES = ['primary', 'secondary'];
+export const DEFAULTS = {
+  config: {
+    alignment: ALIGN_CENTER,
+    size: 'small',
+    width: 'fit-content',
+  },
+  button: {
+    activeButton: 0,
+    borderRadius: 0,
+    borderWidth: 0,
+    padding: 12,
+    buttonText: 'Click Me',
+    url: '',
+  },
+};